+++ src/objects/class_loader.cpp
@@ -82,10 +82,10 @@ static optional<pair<u1 *, size_t>> readClass(const char *path,
             nullptr, 0, nullptr, 0);
 
     auto bytecode = (u1 *) g_heap.allocBytecode(file_info.uncompressed_size); //new u1[file_info.uncompressed_size];
-    int size = unzReadCurrentFile(module_file, bytecode, (unsigned int)(file_info.uncompressed_size));
+    int size = unzReadCurrentFile(module_file, bytecode, (unsigned int) (file_info.uncompressed_size));
     unzCloseCurrentFile(module_file);
     unzClose(module_file);
-    if (size != file_info.uncompressed_size) {
+    if (size != (int) file_info.uncompressed_size) {
         thread_throw(new IOException(NEW_MSG("unzReadCurrentFile failed: %s.\n", path)));
     }
     return make_pair(bytecode, file_info.uncompressed_size);
